changed folder structure for setup and updated moving averages strategy for testing
