+++ losttraderbot/data.py
@@ -11,7 +11,7 @@ import os, os.path
 import numpy as np
 import pandas as pd
 
-from event import Event, MarketEvent
+from .event import Event, MarketEvent
 
 
 class DataHandler(ABC):
