+++ losttraderbot/strategy.py
@@ -10,7 +10,7 @@ import queue
 import numpy as np
 import pandas as pd
 
-from event import SignalEvent
+from .event import SignalEvent
 
 
 class Strategy(ABC):
