+++ setup.py
@@ -13,13 +13,15 @@ setup(
     description="Event-driven Backend Tester",
     long_description=readme,
     long_description_content_type="text/markdown",
-    packages=find_packages(exclude=("examples",)),
+    packages=["losttraderbot"],
     include_package_data=True,
     python_requires=">3.8.*",
     license="MIT License",
     zip_safe=False,
     classifiers=[
+        "Development Statys :: 3 - Alpha",
         "License :: OSI Approved :: MIT License",
+        "Operating System :: OS Independent",
         "Intended Audience :: Developers",
         "Programming Language :: Python :: 3.8",
     ],
